tests/wpt/tests/signed-exchange/sxg-utf8-inner-url.tentative.html

<!DOCTYPE html>
<title>SignedHTTPExchange with UTF-8 inner URL</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  const sxgUrl = get_host_info().HTTPS_ORIGIN + '/signed-exchange/resources/sxg/sxg-utf8-inner-url.sxg';
  const message = await openSXGInIframeAndWaitForMessage(t, sxgUrl);
  assert_equals(message.location, new URL(innerURLOrigin() + '/signed-exchange/resources/\uD83C\uDF10\uD83D\uDCE6.html').href);
  assert_false(message.is_fallback);
}, 'SignedHTTPExchange with UTF-8 inner URL');

</script>
</body>
